BOOKING-163: Moved capacity options into utility file
