+++ assets/src/util/filter-utils.js
@@ -1,0 +1,10 @@
+const capacityOptions = [
+  { value: "0", label: "Alle", type: "gt" },
+  { value: "0..10", label: "0 - 10", type: "between" },
+  { value: "11..20", label: "11 - 20", type: "between" },
+  { value: "21..30", label: "21 - 30", type: "between" },
+  { value: "31..80", label: "31 - 80", type: "between" },
+  { value: "81", label: "81+", type: "gt" },
+];
+
+export default capacityOptions;
